+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The project follows semantic versioning where possible.
 
 ---
 
+## [1.X.X] - 2026-06-05
+- NEW: Blueprint Log Monitor - Show Domoticz log entries; Select log level status, detail or error.
+- NEW: Blueprint Selector Switch - Select level from entries defined in index.html.
+- UPD: Blueprint HMITiles Workbench - Reworked; Dummy Idx and Data.
+- UPD: hmitiles.js - DEBUG flag
+
 ## [1.0.0] - 2026-06-04
 - NEW: Initial public release of HMITiles for Domoticz; Published on the [Domoticz forum](https://forum.domoticz.com/viewtopic.php?t=44682)